requirements.txt
Flask
waitress
Faker
google-cloud-storage
google-genai
google-cloud-bigquery